Move socket functionality to extensions
