+++ hydrus/app.py
@@ -18,7 +18,7 @@ from hydra_python_core import doc_maker
 from hydrus.utils import (
     set_session, set_doc, set_hydrus_server_url,
     set_token, set_api_name, set_authentication)
-from hydrus.socketio_factory import create_socket
+from hydrus.extensions.socketio_factory import create_socket
 
 logger = logging.getLogger(__file__)
